+++ client/src/constants.js
@@ -67,7 +67,6 @@ export const initState = {
 	tournament: {
 		players: null,
 		matches: null,
-		standings: null,
 		pointsToWin: 5,
 		numberOfGames: 1,
 		currentMatchIndex: 0
